requirements.txt
azure_storage==0.37.0
Flask==2.0.3
Flask_Login==0.5.0
Flask_Mail==0.9.1
flask_security==3.0.0
Flask_Security_Too==4.1.3
Flask_SQLAlchemy==2.5.1
msrest==0.6.21
oauthlib==3.2.0
requests==2.27.1
sympy==1.10.1
Werkzeug==2.0.3
